Failed to detect os name on an 32-bit windows machine during "vuls.exe scan" execution
During scanning an 32-bit windows machine, "Failed to detect os name." error occured. By investigation, it is found that formatArch function is called twice during scanning. When first time it is called, it converts the os arch string to "32-bit" , but when it is called second time, with input string 32-bit, it returns "unknown architecture" as there is no switch case to handle "32-bit" case. formatArch function is updated to match proper "x86" or "32-bit" switch case and make the scan to progress and successful.